--- src/components/Home/CorporateContactSection/CorporateContactSection.js
@@ -1,25 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-import FormCard from './FormCard'
-
-export default class CorporateContactSection extends React.Component {
-    render() {
-        return (
-            <div className="centred-container">
-                <h2>Partner with us to <span style={{ color: '#7ac143' }}>make a difference</span></h2>
-                <div className="flex-grid">
-                    <div className="flex-container">
-                        <div className="text-content">
-                            <p>Corporate partnerships will enable us to... Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.</p>
-                        </div>
-                    </div>
-                    <div className="flex-container">
-                        <div className="form-wrapper">
-                            <FormCard></FormCard>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        )
-    }
-}
