--- src/components/Home/ImpactSection/ImpactSection.js
@@ -1,50 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-import CountUp from 'react-countup';
-import VisibilitySensor from 'react-visibility-sensor';
-
-export default class ImpactSection extends React.Component {
-    render() {
-        return (
-            <div className="centred-container">
-                <div className="container">
-                    <h2>Our impact</h2>
-                    <div className="flex-grid">
-                        <div className="flex-container">
-                            <h1 ><CountUp end={983} separator="," redraw={true}>
-                                {({ countUpRef, start }) => (
-                                    <VisibilitySensor onChange={start} delayedCall>
-                                        <span ref={countUpRef} />
-                                    </VisibilitySensor>
-                                )}
-                            </CountUp></h1>
-                            <h3 className="legend-header">Donations</h3>
-                        </div>
-                        <div className="flex-container">
-                            <h1><CountUp end={18652} separator="," redraw={true}>
-                            {({ countUpRef, start }) => (
-                                    <VisibilitySensor onChange={start} delayedCall>
-                                        <span ref={countUpRef} />
-                                    </VisibilitySensor>
-                                )}
-                            </CountUp></h1>
-                            <h3 className="legend-header">Meals Served</h3>
-                        </div>
-                        <div className="flex-container">
-                            <h1><CountUp end={163985} separator="," redraw={true} prefix="$">
-                            {({ countUpRef, start }) => (
-                                    <VisibilitySensor onChange={start} delayedCall>
-                                        <span ref={countUpRef} />
-                                    </VisibilitySensor>
-                                )}
-                            </CountUp></h1>
-                            <h3 className="legend-header">Funds Raised</h3>
-                        </div>
-                    </div>
-                    <p>We are active across in <em>3 states</em>, providing for <em>19 hospitals</em> with <em>#PLACEHOLDER NUMBER# restaurants</em> on board.</p>
-                    <a href="/donation">Get involved</a>
-                </div>
-            </div>
-        )
-    }
-}
