--- src/components/Home/TestimonialSection/TestimonialSection.js
@@ -1,13 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-export default class TestimonialSection extends React.Component {
-    render() {
-        return (
-            <div className="centred-container">
-                <img alt="A health professional" />
-                <blockquote>Insert quote here</blockquote>
-            </div>
-        )
-    }
-}
