--- src/components/Home/ThankYouSponsorsSection/ThankYouSponsorsSection.js
@@ -1,26 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-export default class ThankYouSponsorsSection extends React.Component {
-    render() {
-        return (
-            <div className="centred-container">
-                <h2>Thank you to our partners and sponsors</h2>
-                <div className="flex-grid">
-                    <div className="flex-container">
-                        <img src="" alt="Logo of Company" />
-                    </div>
-                    <div className="flex-container">
-                        <img src="" alt="Logo of Company" />
-                    </div>
-                    <div className="flex-container">
-                        <img src="" alt="Logo of Company" />
-                    </div>
-                    <div className="flex-container">
-                        <img src="" alt="Logo of Company" />
-                    </div>
-                </div>
-            </div>
-        )
-    }
-}
